Disconnected Support: Provide an option to set a custom registry server
In disconnected environments, there is no access to internet and to quay.io, and users should mirror the required images to their internal registry server.
A new parameter named REGISTRY_SERVER is added and if specified, the images will be pulled from it. the relative path of the images will stay the same since they have been mirrored with the oc-mirror tool.

Signed-off-by: Oren Cohen <ocohen@redhat.com>