CNV-71544: Skip .dry_run folder when archiving
when running the checkup with DRY_RUN=true, there are cases in which the folder .dry_run which is used for the progress_watcher, is also archived before it got deleted.
This PR fixes the situation by stopping the progress watcher process before starting to summarizing the results. Also, the junit_parser will skip that folder.

Signed-off-by: Oren Cohen <ocohen@redhat.com>